requirements.txt
albumentations==1.2.1
appdirs==1.4.4
arrow==1.2.2
attrs==21.4.0
Automat==20.2.0
beniget==0.4.1
braceexpand==0.1.7
Brotli==1.0.9
cached-property==1.5.2
Cerberus==1.3.4
certifi-system-store==3021.3.16
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
crit==0.0.1
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.30
distlib==0.3.5
docutils==0.19
eclean-kernel==2.99.3
editables==0.3
file-magic==0.4.0
filelock==3.7.1
flit_core==3.7.1
fonttools==4.28.5
fqdn==1.5.1
gast==0.5.3
gemato==16.2
gentoolkit==0.5.1.post1
gpg==1.17.1
graphviz==0.20
grpcio==1.43.0
h5py==3.7.0
hatch-vcs==0.2.0
hatchling==1.6.0
hyperlink==21.0.0
idna==3.3
imageio==2.20.0
importlib_metadata==4.12.0
incremental==21.3.0
installer==0.5.1
iotop==0.6
iso8601==1.0.2
isoduration==20.11.0
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
java-config==2.3.1
Jinja2==3.1.2
joblib==1.1.0
jsonpointer==2.3
jsonschema==4.7.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lark==1.1.2
layman==2.4.3
lxml==4.9.1
lz4==4.0.1
M2Crypto==0.38.0
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.1
meson==0.62.2
mirrorselect==2.2.6.post2
more-itertools==8.13.0
netifaces==0.11.0
netlink==1.0
networkx==2.8.5
nspektr==0.4.0
numpy==1.23.1
olefile==0.46
opencv-python-headless==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pathspec==0.9.0
pbr==5.9.0
pexpect==4.8.0
Pillow==9.2.0
pipenv==2022.6.7
pkgconfig==1.5.5
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
poetry-core==1.0.8
portage==3.0.30
protobuf==3.19.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pydot==1.4.2
pyelftools==0.28
Pygments==2.12.0
PyGObject==3.42.1
PyHamcrest==2.0.3
pyinotify==0.9.6
pymountboot==0.2.3
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt-builder==1.12.2
PyQt5==5.15.6
PyQt5-sip==12.10.1
pyrsistent==0.18.1
PySocks==1.7.1
pysol-cards==0.14.2
python-dateutil==2.8.2
pythran==0.11.0
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
random2==1.0.1
regex==2022.7.9
rencode==1.0.6
repoman==3.0.3
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
scikit-image==0.19.3
scikit-learn==1.0.2
SciPy==1.8.1
SCons==4.3.0
seccomp==2.5.4
selinux==3.4
semantic-version==2.10.0
service-identity==21.1.0
setuptools-rust==1.4.1
setuptools-scm==7.0.5
sip==6.6.2
six==1.16.0
ssl-fetch==0.4
stevedore==4.0.0
strict-rfc3339==0.7
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2022.5.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
Twisted==22.4.0
typing-extensions==3.10.0.2
uri-template==1.2.0
urllib3==1.26.11
virtualenv==20.16.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webcolors==1.12
webdataset==0.1.96
zipp==3.8.1
zope.interface==5.4.0
